Move tests that require ScanCode to be installed to `ScanCodeScannerFunTest`
Unit tests should not have external dependencies, like the licenses in
the installation directory of ScanCode.

While at it, also inline the `getText()` function to avoid the need to
duplicate it.
